Remove lazy stack frames
